+++ SJF.cpp
@@ -20,13 +20,13 @@ void SJF::get_input_burst_and_arrival_time(QQVector<process> &sjf_process, int n
 
 void SJF::calc_new_order_np(QQVector<process> &sjf_process, QQVector<process> &sjf_process_new, int num_of_proceses)
 {
-    int current_time = 0, min_index = -1;
+    float current_time = 0, min_index = -1;
     // sort wrt arrival time
     sort(sjf_process.begin(), sjf_process.end(), [](process a, process b) { return a.arrival_time < b.arrival_time; });
     current_time = sjf_process[0].arrival_time;
     while (!sjf_process.empty())
     {
-        int min_burst_time = 0, min_index = -1;
+        float min_burst_time = 0, min_index = -1;
         for (size_t i = 0; i < sjf_process.size(); i++)
         {
             if (sjf_process[i].arrival_time <= current_time)
@@ -60,13 +60,13 @@ void SJF::calc_new_order_np(QQVector<process> &sjf_process, QQVector<process> &s
 
 void SJF::calc_new_order_p(QQVector<process> &sjf_process, QQVector<process> &sjf_process_new, int num_of_proceses)
 {
-    int current_time = 0, min_index = -1;
+    float current_time = 0, min_index = -1;
     // sort wrt arrival time
     sort(sjf_process.begin(), sjf_process.end(), [](process a, process b) { return a.arrival_time < b.arrival_time; });
     current_time = sjf_process[0].arrival_time;
     while (!sjf_process.empty())
     {
-        int min_burst_time = 0, min_index = -1;
+        float min_burst_time = 0, min_index = -1;
         for (size_t i = 0; i < sjf_process.size(); i++)
         {
             if (sjf_process[i].arrival_time <= current_time)
@@ -93,7 +93,7 @@ void SJF::calc_new_order_p(QQVector<process> &sjf_process, QQVector<process> &sj
             process current = sjf_process[min_index];
             current.start_time = current_time;
             current_time += sjf_process[min_index].burst_time;
-            int min_burst_time_p = current.burst_time, min_index_p = -1;
+            float min_burst_time_p = current.burst_time, min_index_p = -1;
             for (size_t i = 0; i < sjf_process.size(); i++)
             {
                 if (sjf_process[i].arrival_time <= current_time && i != min_index)
